refactor(comparateur): déplace les questions principales et les réponses éditorialisées dans un sous-dossier et des fichiers distincts
