+++ site/source/pages/simulateurs/comparaison-statuts/simulationConfig.ts
@@ -4,43 +4,16 @@ import { AssimiléSalariéContexte } from '@/domaine/AssimiléSalariéContexte'
 import { ComparateurConfig } from '@/domaine/ComparateurConfig'
 import { IndépendantContexte } from '@/domaine/IndépendantContexte'
 import { AutoEntrepreneurContexteDansPublicodes } from '@/domaine/publicodes/AutoEntrepreneurContexteDansPublicodes'
-import { Question } from '@/domaine/SimulationConfig'
 
 import {
-	réponseActivité,
-	réponseFoyerFiscal,
-	réponseImpôt,
-} from './reponsesEditorialisees'
-
-const questionActivitéPrincipale = {
-	libellé: (t: TFunction) =>
-		t(
-			'pages.simulateurs.comparaison-statuts.questions.activité.principale',
-			'Activité principale'
-		),
-	dottedName: 'entreprise . activité . nature',
-} satisfies Question
-const questionTypeActivité = {
-	libellé: (t: TFunction) =>
-		t(
-			'pages.simulateurs.comparaison-statuts.questions.activité.type',
-			'Type d’activité'
-		),
-	dottedName: 'entreprise . activités . service ou vente',
-} satisfies Question
-const questionActivitéRéglementée = {
-	libellé: (t: TFunction) =>
-		t(
-			'pages.simulateurs.comparaison-statuts.questions.activité.réglementée',
-			'Activité réglementée'
-		),
-	dottedName: 'entreprise . activité . nature . libérale . réglementée',
-} satisfies Question
-const questionAcre = {
-	libellé: (t: TFunction) =>
-		t('pages.simulateurs.comparaison-statuts.questions.acre', 'Acre'),
-	dottedName: 'dirigeant . exonérations . ACRE',
-} satisfies Question
+	questionAcre,
+	questionActivitéPrincipale,
+	questionActivitéRéglementée,
+	questionTypeActivité,
+} from './questions/questionsPrincipales'
+import { réponseActivité } from './questions/réponseActivité'
+import { réponseFoyerFiscal } from './questions/réponseFoyerFiscal'
+import { réponseImpôt } from './questions/réponseImpôt'
 
 export const configComparateurStatuts: ComparateurConfig = {
 	nomModèle: 'modele-social',
